refactoring = HttpHelperAsync non-blocking request issuer
