+++ ExchangeRateProvider/ExchangeRateProvider/ExchangeRateProvider.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Infrastructure\ApiProxy\ApiErrorOccurredEventArgs.cs" />
     <Compile Include="Infrastructure\ApiProxy\ApiProxy.cs" />
     <Compile Include="Infrastructure\HttpHelper\HttpHelper.cs" />
+    <Compile Include="Infrastructure\HttpHelper\HttpHelperAsync.cs" />
     <Compile Include="Infrastructure\HttpHelper\IHttpHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
